% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugify-package.R
\docType{package}
\name{slugify}
\alias{slugify}
\alias{slugify-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/slugify}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/slugify/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
