% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{slugify_fast}
\alias{slugify_fast}
\title{Much faster string slugifier}
\usage{
slugify_fast(x, repl = "-", lower = TRUE)
}
\arguments{
\item{x}{string to slugify}

\item{repl}{what to replace whitespace with; defaults to \verb{'}}

\item{lower}{turn slugified string into lowercase? defaults to \code{TRUE}}
}
\value{
string
}
\description{
Much faster string slugifier
}
\examples{
slugify_basic("R is great!")
}
