% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugify-r.R
\name{slugify_native}
\alias{slugify_native}
\title{Native R slugify (with the help of {stringi})}
\usage{
slugify_native(x, repl = "-", lower = TRUE)
}
\arguments{
\item{x}{string to slugify}

\item{repl}{what to replace spaces with}

\item{lower}{lowercase final output?}
}
\description{
Native R slugify (with the help of {stringi})
}
