% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugify.R
\name{slugify}
\alias{slugify}
\title{Slugify a string to remove characters that can cause problems}
\usage{
slugify(x, alphanum_replace = "", space_replace = "_", tolower = TRUE)
}
\arguments{
\item{x}{a string}

\item{alphanum_replace}{character to replace non-alphanumeric characters}

\item{space_replace}{character to replace spaces}

\item{tolower}{boolean, run towlower()?}
}
\value{
a string with replacements
}
\description{
Slugify a string to remove characters that can cause problems
}
\examples{
slugify("HELLO WORLD!!!")

}
\concept{
slugify
}
