% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugify-package.R, R/slugify-r.R
\docType{package}
\name{slugify}
\alias{slugify-package}
\alias{slugify}
\title{Create Slug Strings for SEO}
\usage{
slugify(x, repl = "-", lower = TRUE)
}
\arguments{
\item{x}{string to slugify}

\item{repl}{what to replace spaces with}

\item{lower}{lowercase final output?}
}
\description{
Most blogging platforms turn title phrases into slugs
for better SEO. Tools are provided to turn phrases into slugs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/slugify}
  \item Report bugs at \url{https://github.com/hrbrmstr/slugify/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
