% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugify-package.R, R/slugify.R
\docType{package}
\name{slugify}
\alias{slugify}
\alias{slugify-package}
\title{Create slug strings for SEO}
\usage{
slugify(x, replacement = "-", remove = NULL, lower = TRUE, strict = TRUE)
}
\arguments{
\item{x}{string}

\item{replacement}{replace spaces with replacement character, defaults to \code{-}}

\item{remove}{remove characters that match this regex, defaults to \code{NULL} (no deleting)}

\item{lower}{convert to lower case, defaults to \code{TRUE}}

\item{strict}{strip special characters except replacement, defaults to \code{TRUE}}
}
\description{
Most blogging platforms turn title phrases into slugs
for better SEO. Tools are provided to turn phrases into slugs.
}
\examples{
slugify("R is great!")
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/slugify}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/slugify/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
